db: fix cancel request race condition on context timeout/cancellation
When passing a timeout context into the db queries, there currently
exists a race condition where a select between query finish and context
done can choose the context, and go on to send a cancel request on a
conn that has already been freed. So a concurrent request may pick up
the same conn but is cancelled by this orphaned request.

There is a similar fix in https://github.com/lib/pq/pull/578 where we
use the same channel to pass flow between fn finish and cancel
goroutines. close() is async so a select cannot guarantee executing one
case at a time.

Added a test in race_test to prove the problem and the fix.