Added support for omitting zero values from batch updates.
In a batch update it already correctly asserts that a zero value field
should be represented as such. But because it sends all data in the
update during a batch update it can still overwrite the existing value.

This solves that problem by adding a COALESCE around the value set
portion of the update statement so that if the value provided from the
ORM is null then it will not overwrite the existing database value.