Added handling of serial column types in updates.
When performing a bulk update previously it would try to cast the fields
that were "bigserial" or "serial" to those types in the from clause. But
because they are not actual types in PostgreSQL and are just aliases, it
would cause an error when executing the query. This resolves that by
making sure that serial columns are cast to their base type instead.

Fields in batch updates will always be typecast.
This is because when a field is not typecast PostgreSQL assumes that
type to be text. Which is fine for some use cases. However if you are
sending an update for a numeric column (or any non-text column) and you
do not specify a type but provide a null value. Then in the COALESCE
function when updating non-zero values PostgreSQL will return an error.
Because COALESCE assumes that both arguments are the same type (relative
to the first argument provided).

Fixed ambiguous column name error in bulk update.