Allow for bulk deletes with custom where clause.
Previously if you were to add your own where clause onto a delete
statement. pg would ignore the original clause (where ID in...) and
would instead only use the clause you had provided. This change makes
it so pg will attempt to use both with an AND so that all conditions
must be met for a record to be deleted when deleting from an array of
models.

This will allow more security in a situation where you are restricting
what a single account could do, as well as other use cases.