Fixed issue where pointer uint64 values would panic during inserts.
This was because we were trying to read the Uint value from the
reflection, but we had the pointer, not the actual reflected value. This
fixes the problem by checking to see if we were given a pointer and if
it is it will get the value from the address.