Introduce environment variables to customize running tests
This commit adds the support of some environment variables (namely,
PGHOST, PGPORT, PGUSER, PGDATABASE, and PGSSLMODE) that override
certain options when running tests. The inspiration is taken from
lib/pq tests (https://www.postgresql.org/docs/9.3/libpq-envars.html).
For every option, the corresponding environment variable is first
checked. If it is present, then the value is used; if it is not present
or empty, then we fall back to using the default values.