Fixed bug when joining a non soft delete table with a soft delete table.
When a non soft delete table was the base table the query builder would
panic when it tried to add the soft delete criteria to the join ON
criteria. This was because we were properly evaluating that a soft
delete filter should be added on the join. But we were adding it for the
wrong table and trying to add it for a table where the soft delete
column was nil causing a panic.

This resolves the problem by making the ON filter for a soft delete join
for the actual join/soft deletable table rather than the base table.