When registering a scanner or an appender, only panic if the type has already been explicitly registered with RegisterScanner or RegisterAppender, rather than if it was implicitly set to a default as a side-effect of types.Appender or types.Scanner being called.
This avoids a race where a previous call to types.Appender or
types.Scanner which was invoked as a side-effect assigned a default
Appender/Scanner which then causes a subsequent call to
Register{Appender,Scanner} to panic. This can happen if, for example,
orm.GetTable is invoked before an init() function that calls Register
gets invoked due to unlucky package import ordering, since orm.GetTable
tries to look up the type's Appender/Scanner.

Example:

  // This package gets imported first
  package bar

  type MyStruct struct {
    MyType MyType
  }

  var Table = orm.TableFor(&MyStruct{})

  // This package gets imported second
  package foo

  type MyType struct {}

  func init() {
    // This panics because orm.TableFor called types.Appender and,
    finding none, set a default
    types.RegisterAppender(&MyType{}, ...)
  }