Add password environment variable for customizing running tests (#1996)
In #1429, environment variables were added to read in postgres
connection details for customizing tests. This change does the
same but with a password variable.

It also removes the explicit user and password values from pgOptions,
as Options init() will set it as postgres anyway if not specified.
This also enables env variables to be used instead for those values.